refactor: topnav
