+++ app/pricing-form/index.tsx
@@ -1,7 +1,6 @@
 'use client'
 
 import React from 'react'
-import styles from './styles.module.css'
 import { FormBlockedFallback, useHubspotFormFallback } from '@/components/HubspotFormFallback'
 
 function PricingForm({ portalId, formId, formName }) {
