--- app/pricing-form/styles.module.css
@@ -1,149 +1,0 @@
-.formLabel {
-  font-family: var(--ifm-font-family-work-sans);
-  font-style: normal;
-  font-weight: 500;
-  font-size: 1.1rem;
-  line-height: var(--ifm-font-line-height);
-  letter-spacing: var(--ifm-font-letter-spacing);
-  margin-bottom: 0.7rem;
-  display: block;
-}
-.textInput {
-  font-family: var(--ifm-font-family-work-sans);
-  width: 100%;
-  padding: 0.8rem 1.2rem;
-  font-size: 1.1rem;
-  border: 1px solid var(--ifm-color-gray-dark);
-  background: var(--ifm-color-gray-dark);
-  border-radius: 0.25rem;
-  color: var(--ifm-color-white-grey);
-}
-.textInput:focus {
-  outline: none !important;
-  border-color: var(--ifm-color-primary);
-}
-.textInput::placeholder {
-  font-family: var(--ifm-font-family-work-sans);
-  color: var(--ifm-color-gray-lightest);
-  font-size: 1.1rem;
-  line-height: 1;
-}
-.textInputLine {
-  font-family: var(--ifm-font-family-work-sans);
-  background: transparent;
-  border: 0;
-  outline: none;
-  border-bottom: 1px solid var(--ifm-color-white);
-  font-size: 0.9rem;
-}
-.textInputLine:focus {
-  border-color: var(--ifm-color-primary);
-}
-.submitBtn {
-  font-size: 1.1rem;
-  line-height: var(--ifm-font-line-height);
-  letter-spacing: var(--ifm-font-letter-spacing);
-  font-family: var(--ifm-font-family-work-sans);
-  width: 100%;
-  padding: 0.8rem 1.2rem;
-  border: 0;
-  border-radius: 0.25rem;
-  font-weight: 600;
-}
-.inputGroup {
-  margin-bottom: 1.6rem;
-}
-.formCheckWrapper {
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
-  margin-bottom: 0.2rem;
-}
-.formCheckWrapper label {
-  font-size: 0.9rem;
-  font-family: var(--ifm-font-family-work-sans);
-  line-height: var(--ifm-font-line-height);
-}
-.checkInput {
-  font-family: var(--ifm-font-family-work-sans);
-  border: 0;
-  background: var(--ifm-color-gray-dark);
-  border-radius: 0.25rem;
-  color: var(--ifm-color-white-grey);
-  width: 1.1rem;
-  height: 1.1rem;
-  margin-right: 0.7rem;
-  accent-color: var(--ifm-color-primary);
-}
-.radioInput {
-  font-family: var(--ifm-font-family-work-sans);
-  border: 0;
-  background: var(--ifm-color-gray-dark);
-  border-radius: 0.25rem;
-  color: var(--ifm-color-white-grey);
-  width: 1.1rem;
-  height: 1.1rem;
-  margin-right: 0.3rem;
-  accent-color: var(--ifm-color-primary);
-}
-.formRadioWrapper {
-  display: inline-flex;
-  justify-content: center;
-  align-items: center;
-  margin-right: 0.5rem;
-}
-.formRadioWrapper label {
-  font-size: 0.9rem;
-  font-family: var(--ifm-font-family-work-sans);
-  line-height: var(--ifm-font-line-height);
-}
-.error {
-  color: var(--ifm-color-primary);
-  font-size: 0.8rem;
-}
-
-@media (max-width: 768px) {
-  .formLabel {
-    font-size: 0.9rem;
-  }
-  .textInput {
-    padding: 0.5rem 0.7rem;
-    font-size: 0.8rem;
-  }
-  .textInput::placeholder {
-    font-size: 0.8rem;
-  }
-  .submitBtn {
-    font-size: 0.8rem;
-    padding: 0.5rem 0.7rem;
-  }
-  .formCheckWrapper label {
-    font-size: 0.8rem;
-  }
-  .textInputLine {
-    font-size: 0.8rem;
-    width: 50%;
-  }
-  .formRadioWrapper label {
-    font-size: 0.8rem;
-  }
-  .checkInput,
-  .radioInput {
-    width: 0.75rem;
-    height: 0.75rem;
-  }
-}
-html[data-theme="light"] .textInput {
-  background: var(--ifm-color-white);
-  border: 1px solid var(--ifm-color-gray-dark);
-  color: var(--ifm-color-gray-darkest);
-}
-html[data-theme="light"] .textInputLine {
-  border-bottom: 1px solid var(--ifm-color-gray-dark);
-}
-html[data-theme="light"] .textInput:focus {
-  border-color: var(--ifm-color-primary);
-}
-html[data-theme="light"] .textInputLine:focus {
-  border-color: var(--ifm-color-primary);
-}
