--- components/TopNav/styles.module.css
@@ -1,34 +1,0 @@
-.topnavItemDropdownMenu {
-  width: auto !important;
-}
-
-.topnavItemDropdownMenu ul {
-  display: flex;
-  flex-direction: row;
-  gap: 16px;
-
-  width: 600px;
-}
-
-.modulesSection {
-  padding: 0.5rem 0;
-  width: 60%;
-}
-
-.modulesSection ul {
-  display: flex !important;
-  flex-direction: column !important;
-  gap: 16px !important;
-}
-
-.successStoriesSection {
-  padding: 0.5rem 0;
-
-  width: 40%;
-}
-
-.successStoriesSection ul {
-  display: flex !important;
-  flex-direction: column !important;
-  gap: 16px !important;
-}
